pyproject.toml
[tool.poetry]
name = "schemainspect"
version = "3.1"
authors = ["Robert Lechte <robertlechte@gmail.com>"]
license = "Unlicense"
readme = "README.md"
description = "Schema inspection for PostgreSQL (and possibly others)"

repository = "https://github.com/djrobstep/schemainspect"
homepage = "https://github.com/djrobstep/schemainspect"

[tool.poetry.dependencies]
python = ">=3.6.2 <4"
sqlalchemy = "*"
PyYAML = "^6.0"
pgpasslib = "^1.1.0"

[tool.poetry.dev-dependencies]
sqlbag = ">=0.1.1616028516"
pytest = {version="*", python=">=3.5 <4"}
pytest-cov = "*"
pytest-clarity = {version=">=0.3.0-alpha.0", python=">=3.5 <4"}
psycopg2-binary = "*"
flake8 = "*"
isort = {version=">=5", python=">=3.6 <4"}
migra = "*"
black = "22.3.0"
toml = "*"

[tool.poetry.scripts]
schemainspect = 'schemainspect:do_command'

[tool.isort]
profile = "black"
